Change Return Decimals Value.

+++ contracts/TetherToken.sol
@@ -41,7 +41,7 @@ contract TetherToken is Context, IERC20USDTCOMPATIBLE {
         return _symbol;
     }
 
-    function decimals() public view override returns (uint256) {
+    function decimals() public view override returns (uint8) {
         return _decimals;
     }
